Split long lines

+++ contracts/Escrow.sol
@@ -288,7 +288,11 @@ contract Escrow is IERC777Recipient, IERC777Sender, IEscrow, Permissions {
      *
      * - Beneficiary and Vesting manager must be `msg.sender`.
      */
-    function cancelPendingDelegation(uint delegationId) external override onlyActiveBeneficiaryOrVestingManager {
+    function cancelPendingDelegation(uint256 delegationId)
+        external
+        override
+        onlyActiveBeneficiaryOrVestingManager
+    {
         IDelegationController delegationController = IDelegationController(
             contractManager.getContract("DelegationController")
         );
